refactor: domain, service, repository 패키지 생성 후 분리
