+++ BE/src/main/java/team25/issuetracker/controller/OauthRestController.java
@@ -1,4 +1,4 @@
-package team25.issuetracker;
+package team25.issuetracker.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import team25.issuetracker.LoginResponse;
+import team25.issuetracker.service.OauthService;
 
 @Slf4j
 @RestController
