+++ BE/src/main/java/team25/issuetracker/repository/MemberRepository.java
@@ -1,7 +1,8 @@
-package team25.issuetracker;
+package team25.issuetracker.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import team25.issuetracker.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
